feat(store): add versioned serialization/deserialization
Forward compatibility means the encoded V is of a newer version than
what our software has, but our software needs to support deserializing
it.
To have forward compatibility, we need a length prefix for every encoded
V.

Why? Typically, the encoding length can be determined by the type
definition of V. However, a newer version of V will have new variants,
and the encoding lengths of those variants cannot be determined (because
we don't have it).
When we fail to deserialize into a variable of type V, we need to check
the "variant index" encoding first.

Why? If the variant index does not exist in our type V (which is an
enum), then we are sure the data if of a newer type. Since the data is
of a newer type, we can apply the forward-compatibility logic. On the
flip side, if the data is of a variant index that exists in V, then we
are sure that the data is corrupted.

What is "variant index"? bincode encodes enums with the variant index
first. The variant index is encoded with VarintEncoding by default.

ENCODING FORMAT

Each changeset should be encoded as follows:

[L][V]

- [L] is a bincode encoded u64. Use serialized_size on V to figure out
  the value of L.
- [V] is a bincode encoded enum type that represents a versioned
  changeset. Because it is an enum, it is guaranteed that the encoded
  data begins with a VariantEncoding.

Encoding:

1. Transform the changeset into a versioned-changeset first.
2. Determine the serialized_size of the versioned-changeset. The
   serialized_size of V is L.
3. Encode L: u64 and write to file.
4. Encode V and write to file.

Decoding:

1. Decode L: u64 from file.
2. Read L bytes from file and attempt to decode as V.
  a. On success: Great! Extract C from V and return it.
  b. On failure: Decode V as u64 (to extract the
  enum-variant-varint-index). If variant index is too high, trim the
  variant-index from the data, and attempt to decode into C (changeset)
  directly.

How do we determine if a variant-index is too high?

Encode V::default() (which should always be the latest-version variant),
and then decode the result as u64 to extract the variant-index.