feat(store): add parse logic for variants with extra fields
The V generic is intended for cases when the user keeps their changesets
versioned and with a strict policy of only adding fields, not modifying
older ones nor removing them.

In the case the data has been encoded with a newer version, but is
decoded with an older one, the code should attemp to parse only the
fields it already knows of, ignoring the rest.

This code implements that logic but also opens the door for parsing
fields of different types as the wrong type.