Depend on a new psbt-v0 crate
I've hacked up a new `psbt-v0` crate with the goals of:

- Create crate directly from `rust-bitcoin v0.32.2` and `rust-miniscript
  12.0.0` with minimal changes so it can be diff'ed an verified.
- Be a drop in replacement for the `psbt` modules in `bitcoin` and `miniscript`.

This hopes to enable the following:

- Make development on PSBT separate from development on `rust-bitcoin`
- Get a different set of maintainers involved
- Make development of PSBT crate more rapid, more fun, and more productive

My claim is that PSBT has different stakeholders to other code in
`bitcoin` and `miniscript` and the fact that it lives in these two
crates (specifically `bitcoin`) makes it harder to attract devs and
harder to patch than necessary.

github: https://github.com/tcharding/rust-psbt-v0

Potential Roadmap:

`v0.1.0` would be the drop in replacement with minimal changes. Next
we could perhaps do:

- Re-write signing API
- Implement PSBTv2
- Re-write the whole thing with what we've learned over the last few
  years