ref(wallet): move getting local utxos to `create_tx`
We can't easily plan the utxos before all of the assets are known,
so having `TxBuilder` try to create `WeightedUtxo`s creates a
chicken and egg problem that we fix by moving some logic to
`Wallet::create_tx`.

`TxParams::utxos` field is thus split into two: `utxos` is a list
of outpoints strictly to be used for getting locally owned outputs,
and `foreign_utxos` handles adding `WeightedUtxos` coming from
`add_foreign_utxo`. These are later combined with the manually
selected coins to form the list of utxos that must be spent.