feat(wallet)!: use miniscript `plan` in place of old policy mod
The internal behavior is preserved insofar as the policy
path can influence the outcome of tx building. The difference
is instead of encode the path as nodes in a policy tree, the
plan is determined by the `Assets` available.

Note that creating a `Plan` is a pre-requisite to coin selection,
so if a UTXO in the wallet can't be planned, it won't be selected.
The wallet attempts to create a plan for each UTXO based on the
available information. The user may provide additional assets or
specify a particular set of assets in order to complete a plan.

New APIs include `TxBuilder::add_assets` and `Wallet::try_plan`
and `Wallet::planned_utxos`.

To prevent confusion due to overlapping functionality, the
`TxBuilder` methods `policy_path` and `current_height` are
removed.