feat(wallet): `create_single` accepts a multipath descriptor
Test that creating a Wallet from a multipath descriptor returns a
Wallet with two keychains. The first part of the multipath goes to
the first (External) keychain and the second part goes to the
Internal keychain.

Fixup error message for `DescriptorError::Multipath` which
previously indicated they weren't supported. The error can
still be triggered by the appearance of a MultiXPub where
not explicitly allowed.