fix(core): Memory leak bugs in `CheckPoint::drop` impl
Fix memory leak bug in `CheckPoint::drop` by using `Arc::into_inner` if
it is available (>= 1.70).

Fix `CPInner::drop` logic so that if `CPInner::block` becomes
generic and is of a type that required `drop`, it does not leak memory.

Add tests for memory leak + stack overflow when dropping `CheckPoint`.