refactor(bip158)!: make constructor `FilterIter::new_with_checkpoint` fallible
The logic of `find_base` is moved inside the `new_with_checkpoint`
constructor, which will error if a point of agreement (PoA) isn't
found. As a result, `get_tip` is now only responsible for setting
the stop height, which means it can be called multiple times on
a single instance of FilterIter.

`cp` field is removed from FilterIter struct, since it is now
only used in the constructor to find a base and set the start
height.

There's no longer a need to remember conflicting blocks on the way
to finding the PoA, now that we store all newly fetched headers
during `next`.