feat: add utilities for testing persistence
Added the following functions to test if persistence of `bdk_chain` is
happening correctly.
  - `persist_txgraph_changeset`
  - `persist_indexer_changeset`
  - `persist_local_chain_changeset`
which leverage a more general `persist_changeset`. In order to not panic
and instead return errors raised by the persistence backend being
tested, introduced `PersistErr`. To ensure downcasting and to allow
usage of anyhow, `PersisterErr::Persister` requires an error of with
'static + Send + Sync.