feat(chain)!: introduce new `TopologicalIterWithLevels`
- add two new fields to `CanonicalIter`, the `canonical_roots` and
  `canonical_ancestors`, they're used by the `TopologicalIterWithLevels`
in order to know what are the `roots` that it should start emitting
from, and what are the ancestors/parents of each canonical transaction,
in order to build the inputs dependencies and parents count.
- add new `TopologicalIterWithLevels`, it's used to topologically
  iterate through the canonicalized `TxGraph`, emitting each level and
it's canonical descendants in a topological order, following the correct
spending order (e.g. tx1 -> tx2 and tx1, tx2 -> tx3, so it should output
tx1, tx2 and then tx3 as it's spends both tx1 and tx2).